Reorganizing Dynamic Programming Problems - Contribution#398
