+++ Algorithms.Tests/Problems/DynamicProgramming/CoinChange/GenerateChangesDictionaryTests.cs
@@ -1,9 +1,9 @@
 ﻿using System.Linq;
-using Algorithms.Problems.CoinChange;
+using Algorithms.Problems.DynamicProgramming.CoinChange;
 using FluentAssertions;
 using NUnit.Framework;
 
-namespace Algorithms.Tests.Problems.CoinChange.Dynamic
+namespace Algorithms.Tests.Problems.DynamicProgramming.CoinChange
 {
     [TestFixture]
     public class GenerateChangesDictionaryTests
