+++ Algorithms.Tests/Problems/DynamicProgramming/CoinChange/GenerateSingleCoinChangesTests.cs
@@ -1,10 +1,10 @@
 ﻿using System;
 using System.Linq;
-using Algorithms.Problems.CoinChange;
+using Algorithms.Problems.DynamicProgramming.CoinChange;
 using FluentAssertions;
 using NUnit.Framework;
 
-namespace Algorithms.Tests.Problems.CoinChange.Dynamic
+namespace Algorithms.Tests.Problems.DynamicProgramming.CoinChange
 {
     [TestFixture]
     public class GenerateSingleCoinChangesTests
